TOUR Header Background Image
I have changed the background image of  TOURS header in TOURS page, i have added the image of a lake which i had already asked to upload as i didn't have access to upload files.
I had to add many different background image css commands to make it more appealing. you can look at the changes in Styles.

A class name col-12 of the division (<div>)was left unused, i've used the same as the container for the image.